improve scripts
